tgnotes/requirements.txt
Babel==0.9.6
Beaker==1.5.4
FormEncode==1.2.4
Genshi==0.6
Mako==0.7.0
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.3.4
PasteScript==1.7.5
Pygments==1.4
Pylons==1.0
Routes==1.12.3
SQLAlchemy==0.7.7
Tempita==0.5.1
ToscaWidgets==0.9.12
TurboGearNotes==0.1dev
TurboGears2==2.1.5
WebError==0.10.3
WebFlash==0.1a9
WebHelpers==1.3
WebOb==1.0.8
WebTest==1.2.4
decorator==3.3.2
docutils==0.9
example==0.1dev
nose==1.0.0
repoze.tm2==1.0
simplejson==2.3.0
sprox==0.7.2
sqlalchemy-migrate==0.7.2
tg.devtools==2.1.5
tgext.crud==0.5
transaction==1.2.0
tw.forms==0.9.9
wsgiref==0.1.2
zope.interface==3.8.0
zope.sqlalchemy==0.7
